Scoreboard components
Refactor the scoreboard into two components: scoreboard header and scoreboard row.

This doesn't change anything in the current UI, and functionally these two
components will always depend on each other and have shared code. So why do this?
Over time this separation would allow:
- scoreboard rows to be shown independently, e.g. on a team's page.
- individual group/region scoreboards.
- different variants of the scoreboard, e.g. paged vs scrolling, showing the
  same header at the bottom, or multi-column.
